Move twisted_thread from Core.Utilities to Test
