+++ Tribler/Test/API/test_download.py
@@ -1,14 +1,16 @@
-from binascii import hexlify
 import logging
 import os
 import shutil
+from binascii import hexlify
 from unittest import skip
+
 from twisted.internet.defer import Deferred
+
 from Tribler.Core.Utilities.network_utils import get_random_port
-from Tribler.Core.Utilities.twisted_thread import deferred
 from Tribler.Core.simpledefs import dlstatus_strings, DLSTATUS_DOWNLOADING
 from Tribler.Test.common import UBUNTU_1504_INFOHASH, TORRENT_UBUNTU_FILE
 from Tribler.Test.test_as_server import TestAsServer
+from Tribler.Test.twisted_thread import deferred
 
 
 class TestDownload(TestAsServer):
