+++ Tribler/Test/API/test_seeding.py
@@ -6,10 +6,10 @@ import os
 from twisted.internet.defer import inlineCallbacks, Deferred
 
 from Tribler.Core.TorrentDef import TorrentDef
-from Tribler.Core.Utilities.twisted_thread import deferred
 from Tribler.Core.simpledefs import DLSTATUS_SEEDING, dlstatus_strings
-from Tribler.Test.test_as_server import TestAsServer
 from Tribler.Test.common import TESTS_DATA_DIR
+from Tribler.Test.test_as_server import TestAsServer
+from Tribler.Test.twisted_thread import deferred
 from Tribler.dispersy.util import blocking_call_on_reactor_thread
