+++ Tribler/Test/Core/Modules/RestApi/Channels/test_channels_discovered_endpoint.py
@@ -1,6 +1,6 @@
 from Tribler.Core.Modules.restapi.channels.base_channels_endpoint import UNKNOWN_CHANNEL_RESPONSE_MSG
-from Tribler.Core.Utilities.twisted_thread import deferred
 from Tribler.Test.Core.Modules.RestApi.Channels.test_channels_endpoint import AbstractTestChannelsEndpoint
+from Tribler.Test.twisted_thread import deferred
 
 
 class TestChannelsDiscoveredEndpoints(AbstractTestChannelsEndpoint):
