+++ Tribler/Test/Core/Modules/RestApi/Channels/test_channels_endpoint.py
@@ -1,12 +1,11 @@
 import json
+
 from Tribler.Core.Modules.channel.channel import ChannelObject
 from Tribler.Core.Modules.channel.channel_manager import ChannelManager
-
-from Tribler.Core.Utilities.twisted_thread import deferred
 from Tribler.Core.exceptions import DuplicateChannelNameError
-from Tribler.Core.simpledefs import NTFY_CHANNELCAST, NTFY_VOTECAST
 from Tribler.Test.Core.Modules.RestApi.base_api_test import AbstractApiTest
 from Tribler.Test.Core.base_test_channel import BaseTestChannel
+from Tribler.Test.twisted_thread import deferred
 
 
 class ChannelCommunityMock(object):
