+++ Tribler/Test/Core/Modules/RestApi/Channels/test_channels_playlist_endpoint.py
@@ -2,9 +2,9 @@ import json
 from twisted.internet.defer import inlineCallbacks
 
 from Tribler.Core.Modules.restapi.channels.base_channels_endpoint import UNKNOWN_CHANNEL_RESPONSE_MSG
-from Tribler.Core.Utilities.twisted_thread import deferred
 from Tribler.Test.Core.Modules.RestApi.Channels.test_channels_endpoint import AbstractTestChannelsEndpoint
 from Tribler.Test.Core.base_test import MockObject
+from Tribler.Test.twisted_thread import deferred
 from Tribler.dispersy.exception import CommunityNotFoundException
