+++ Tribler/Test/Core/Modules/RestApi/Channels/test_channels_subscription_endpoint.py
@@ -1,18 +1,12 @@
 import time
-
-from twisted.internet.defer import inlineCallbacks, succeed
+from twisted.internet.defer import succeed
 
 from Tribler.Core.Modules.restapi import VOTE_SUBSCRIBE, VOTE_UNSUBSCRIBE
 from Tribler.Core.Modules.restapi.channels.base_channels_endpoint import UNKNOWN_CHANNEL_RESPONSE_MSG
 from Tribler.Core.Modules.restapi.channels.channels_subscription_endpoint import ALREADY_SUBSCRIBED_RESPONSE_MSG, \
     NOT_SUBSCRIBED_RESPONSE_MSG
-from Tribler.Core.Utilities.twisted_thread import deferred
 from Tribler.Test.Core.Modules.RestApi.Channels.test_channels_endpoint import AbstractTestChannelsEndpoint
-from Tribler.community.allchannel.community import AllChannelCommunity
-from Tribler.dispersy.dispersy import Dispersy
-from Tribler.dispersy.endpoint import ManualEnpoint
-from Tribler.dispersy.member import DummyMember
-from Tribler.dispersy.util import blocking_call_on_reactor_thread
+from Tribler.Test.twisted_thread import deferred
 
 
 class TestChannelsSubscriptionEndpoint(AbstractTestChannelsEndpoint):
