+++ Tribler/Test/Core/Modules/RestApi/Channels/test_channels_torrents_endpoint.py
@@ -1,17 +1,16 @@
 import base64
 import json
-import urllib
 import os
 import shutil
+import urllib
 from twisted.internet.defer import inlineCallbacks
-
 
 from Tribler.Core.TorrentDef import TorrentDef
 from Tribler.Core.Utilities.network_utils import get_random_port
-from Tribler.Core.Utilities.twisted_thread import deferred
 from Tribler.Test.Core.Modules.RestApi.Channels.test_channels_endpoint import AbstractTestChannelsEndpoint
 from Tribler.Test.Core.base_test import MockObject
 from Tribler.Test.common import TORRENT_UBUNTU_FILE
+from Tribler.Test.twisted_thread import deferred
 from Tribler.dispersy.exception import CommunityNotFoundException
