+++ Tribler/Test/Core/Modules/RestApi/Channels/test_create_channel_endpoint.py
@@ -1,6 +1,7 @@
 import json
-from Tribler.Core.Utilities.twisted_thread import deferred
+
 from Tribler.Test.Core.Modules.RestApi.Channels.test_channels_endpoint import AbstractTestChannelsEndpoint
+from Tribler.Test.twisted_thread import deferred
 
 
 class TestCreateChannelEndpoint(AbstractTestChannelsEndpoint):
