+++ Tribler/Test/Core/Modules/RestApi/base_api_test.py
@@ -1,20 +1,17 @@
 import json
 import os
 import urllib
-
-from zope.interface import implements
-
 from twisted.internet.defer import succeed, inlineCallbacks
 from twisted.python.threadable import isInIOThread
 from twisted.web.client import Agent, readBody, HTTPConnectionPool
 from twisted.web.http_headers import Headers
 from twisted.web.iweb import IBodyProducer
+from zope.interface import implements
 
 from Tribler.Core.Utilities.network_utils import get_random_port
-from Tribler.Core.Utilities.twisted_thread import reactor
 from Tribler.Core.version import version_id
-from Tribler.Test.Core.base_test_channel import BaseTestChannel
 from Tribler.Test.test_as_server import TestAsServer
+from Tribler.Test.twisted_thread import reactor
 from Tribler.dispersy.util import blocking_call_on_reactor_thread
