+++ Tribler/Test/Core/Modules/RestApi/test_create_torrent_endpoint.py
@@ -1,14 +1,13 @@
-import json
 import base64
+import json
 import os
 import shutil
-
 from twisted.internet.defer import inlineCallbacks
 
-from Tribler.Core.Utilities.twisted_thread import deferred
 from Tribler.Core.TorrentDef import TorrentDef
 from Tribler.Test.Core.Modules.RestApi.base_api_test import AbstractApiTest
 from Tribler.Test.common import TESTS_DATA_DIR
+from Tribler.Test.twisted_thread import deferred
 from Tribler.dispersy.util import blocking_call_on_reactor_thread
