+++ Tribler/Test/Core/Modules/RestApi/test_downloads_endpoint.py
@@ -1,13 +1,13 @@
-from binascii import hexlify
 import json
 import os
+from binascii import hexlify
 from urllib import pathname2url
 
 from Tribler.Core.DownloadConfig import DownloadStartupConfig
 from Tribler.Core.Utilities.network_utils import get_random_port
-from Tribler.Core.Utilities.twisted_thread import deferred
 from Tribler.Test.Core.Modules.RestApi.base_api_test import AbstractApiTest
 from Tribler.Test.common import UBUNTU_1504_INFOHASH, TESTS_DATA_DIR
+from Tribler.Test.twisted_thread import deferred
 
 
 class TestDownloadsEndpoint(AbstractApiTest):
