+++ Tribler/Test/Core/Modules/RestApi/test_multichain_endpoint.py
@@ -1,18 +1,18 @@
-import json
 import base64
+import json
 from urllib import quote_plus
 
 from twisted.internet.defer import inlineCallbacks
 
-from Tribler.Core.Utilities.twisted_thread import deferred
-from Tribler.Test.Community.Multichain.test_multichain_utilities import TestBlock
-from Tribler.Test.Core.Modules.RestApi.base_api_test import AbstractApiTest
-from Tribler.Test.Core.base_test import MockObject
 from Tribler.community.multichain.community import MultiChainCommunity
 from Tribler.dispersy.dispersy import Dispersy
 from Tribler.dispersy.endpoint import ManualEnpoint
 from Tribler.dispersy.member import DummyMember
 from Tribler.dispersy.util import blocking_call_on_reactor_thread
+from Tribler.Test.Community.Multichain.test_multichain_utilities import TestBlock
+from Tribler.Test.Core.Modules.RestApi.base_api_test import AbstractApiTest
+from Tribler.Test.Core.base_test import MockObject
+from Tribler.Test.twisted_thread import deferred
 
 
 class TestMultichainStatsEndpoint(AbstractApiTest):
