+++ Tribler/Test/Core/Modules/RestApi/test_rest_manager.py
@@ -1,9 +1,7 @@
 import json
 
-from Tribler.Core.Utilities.twisted_thread import deferred
-
 from Tribler.Core.exceptions import TriblerException
-
+from Tribler.Test.twisted_thread import deferred
 from base_api_test import AbstractApiTest
