+++ Tribler/Test/Core/Modules/RestApi/test_state_endpoint.py
@@ -1,5 +1,5 @@
-from Tribler.Core.Utilities.twisted_thread import deferred
 from Tribler.Test.Core.Modules.RestApi.base_api_test import AbstractApiTest
+from Tribler.Test.twisted_thread import deferred
 
 
 class TestStateEndpoint(AbstractApiTest):
