+++ Tribler/Test/Core/Modules/RestApi/test_statistics_endpoint.py
@@ -1,7 +1,7 @@
 import json
 
-from Tribler.Core.Utilities.twisted_thread import deferred
 from Tribler.Test.Core.Modules.RestApi.base_api_test import AbstractApiTest
+from Tribler.Test.twisted_thread import deferred
 
 
 class TestStatisticsEndpoint(AbstractApiTest):
