+++ Tribler/Test/Core/Modules/RestApi/test_torrents_endpoint.py
@@ -1,13 +1,13 @@
 import json
 import time
 from twisted.internet.defer import inlineCallbacks
+
 from Tribler.Core.TorrentChecker.torrent_checker import TorrentChecker
 from Tribler.Core.Utilities.network_utils import get_random_port
-
-from Tribler.Core.Utilities.twisted_thread import deferred
 from Tribler.Core.simpledefs import NTFY_CHANNELCAST, NTFY_TORRENTS
 from Tribler.Test.Core.Modules.RestApi.base_api_test import AbstractApiTest
 from Tribler.Test.Core.base_test import MockObject
+from Tribler.Test.twisted_thread import deferred
 from Tribler.Test.util.Tracker.HTTPTracker import HTTPTracker
 from Tribler.Test.util.Tracker.UDPTracker import UDPTracker
 from Tribler.dispersy.util import blocking_call_on_reactor_thread
