+++ Tribler/Test/Core/TorrentChecker/test_torrentchecker.py
@@ -1,5 +1,6 @@
 import time
 from twisted.internet.defer import Deferred
+
 from Tribler.Core.CacheDB.SqliteCacheDBHandler import TorrentDBHandler
 from Tribler.Core.Category.Category import Category
 from Tribler.Core.Modules.tracker_manager import TrackerManager
@@ -7,9 +8,9 @@ from Tribler.Core.Session import Session
 from Tribler.Core.SessionConfig import SessionStartupConfig
 from Tribler.Core.TorrentChecker.session import HttpTrackerSession
 from Tribler.Core.TorrentChecker.torrent_checker import TorrentChecker
-from Tribler.Core.Utilities.twisted_thread import deferred
 from Tribler.Core.simpledefs import NTFY_TORRENTS
 from Tribler.Test.Core.base_test import TriblerCoreTest
+from Tribler.Test.twisted_thread import deferred
 from Tribler.dispersy.util import blocking_call_on_reactor_thread
