+++ Tribler/Test/Core/test_network_utils.py
@@ -1,12 +1,11 @@
 import random
 import socket
-
-from Tribler.Core.Utilities.twisted_thread import reactor, deferred
 from nose.tools import raises
 from twisted.internet.protocol import Factory
 
 from Tribler.Core.Utilities.network_utils import get_random_port, autodetect_socket_style, InterruptSocket
 from Tribler.Test.Core.base_test import TriblerCoreTest
+from Tribler.Test.twisted_thread import reactor, deferred
 
 
 class TriblerCoreTestNetworkUtils(TriblerCoreTest):
